move things to package_manager/utils.go
